Update to use new log event
Major changes include removal of xlog/logvent middleware installation.
This is now delegated to consumers of the library. This has two benefits.

One is that this library is now only concerned with general logging structure
and is unopinionated about what kind of underlying logger is in use by
leveraging the logevent abstraction without direct xlog references.

The other benefit is that this also separates log installation from
access logging (and other middleware) such that we'll no longer have
ordering conflicts with tracing.